ai/requirements.txt
transformers
langchain
huggingface_hub
python-dotenv
langchain_mistralai
grpcio 
grpcio-tools
langchain-community
opencv-python 
opencv-python-headless
PyPDF2
docling
docling-core
langchain
groq
langchain_groq
langchain-huggingface
sentence-transformers
bs4
python-dotenv
faiss-cpu
chromadb
fastapi
langgraph
